Build game tree taking into account many moves of one pawn (beating); update game state rate (no opponent); set game state rate while building game tree; add tests; refactorization (#5)

+++ src/main/java/pl/games/checkers/ai/Rate.java
@@ -2,6 +2,6 @@ package pl.games.checkers.ai;
 
 public interface Rate {
 
-	Integer rate();
+	Long rate();
 
 }
